requirements.txt
Flask>=3
Authlib>=1.3
python-dotenv
Werkzeug>=3
gunicorn